magic selecting contiguous is working.
a problem cause slow selection. remove from last of array will be much much faster.
DOC says: On large arrays, this method will be slower if the removed element is close to the beginning of the array (index 0). This is because all elements placed after the removed element have to be reindexed.